[ANDWL-282] currency dialog code change - xml comment delete - itemdecoration class add - watchadd dialog fix - unused code delete
