+++ app/src/main/java/wannabit/io/cosmostaion/dialog/CurrencySetDialog.java
@@ -20,6 +20,7 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import wannabit.io.cosmostaion.R;
+import wannabit.io.cosmostaion.utils.SimpleDividerItemDecoration;
 
 public class CurrencySetDialog extends DialogFragment {
 
@@ -88,32 +89,6 @@ public class CurrencySetDialog extends DialogFragment {
                 super(itemView);
                 rootLayer = itemView.findViewById(R.id.rootLayer);
                 currencyName = itemView.findViewById(R.id.currencyName);
-            }
-        }
-    }
-
-    public static class SimpleDividerItemDecoration extends RecyclerView.ItemDecoration {
-        private Drawable mDivider;
-
-        public SimpleDividerItemDecoration(Context context) {
-            mDivider = ContextCompat.getDrawable(context, R.drawable.box_line_divider);
-        }
-
-        public void onDrawOver(Canvas c, RecyclerView parent, RecyclerView.State state) {
-            int left = parent.getPaddingLeft();
-            int right = parent.getWidth() - parent.getPaddingRight();
-
-            int childCount = parent.getChildCount();
-            for (int i = 0; i < childCount; i++) {
-                View child = parent.getChildAt(i);
-
-                RecyclerView.LayoutParams params = (RecyclerView.LayoutParams) child.getLayoutParams();
-
-                int top = child.getBottom() + params.bottomMargin;
-                int bottom = top + mDivider.getIntrinsicHeight();
-
-                mDivider.setBounds(left, top, right, bottom);
-                mDivider.draw(c);
             }
         }
     }
