+++ app/src/main/java/wannabit/io/cosmostaion/dialog/SelectChainListDialog.java
@@ -90,9 +90,7 @@ public class SelectChainListDialog extends DialogFragment {
         } else {
             mDialogTitle.setText(getSActivity().getString(R.string.str_select_chains));
             mBtnLayer.setVisibility(View.VISIBLE);
-            mBtnLeft.setOnClickListener(View -> {
-                getDialog().dismiss();
-            });
+            mBtnLeft.setOnClickListener(View -> getDialog().dismiss());
             mBtnRight.setOnClickListener(View -> {
                 if (selectedSet.size() >= 1) {
                     mSelectChainsDialogResult.SelectedChains(new Gson().toJson(selectedSet));
