+++ app/src/main/java/wannabit/io/cosmostaion/utils/SimpleDividerItemDecoration.java
@@ -1,0 +1,37 @@
+package wannabit.io.cosmostaion.utils;
+
+import android.content.Context;
+import android.graphics.Canvas;
+import android.graphics.drawable.Drawable;
+import android.view.View;
+
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.RecyclerView;
+
+import wannabit.io.cosmostaion.R;
+
+public class SimpleDividerItemDecoration extends RecyclerView.ItemDecoration {
+    private Drawable mDivider;
+
+    public SimpleDividerItemDecoration(Context context) {
+        mDivider = ContextCompat.getDrawable(context, R.drawable.box_line_divider);
+    }
+
+    public void onDrawOver(Canvas c, RecyclerView parent, RecyclerView.State state) {
+        int left = parent.getPaddingLeft();
+        int right = parent.getWidth() - parent.getPaddingRight();
+
+        int childCount = parent.getChildCount();
+        for (int i = 0; i < childCount; i++) {
+            View child = parent.getChildAt(i);
+
+            RecyclerView.LayoutParams params = (RecyclerView.LayoutParams) child.getLayoutParams();
+
+            int top = child.getBottom() + params.bottomMargin;
+            int bottom = top + mDivider.getIntrinsicHeight();
+
+            mDivider.setBounds(left, top, right, bottom);
+            mDivider.draw(c);
+        }
+    }
+}
