+++ app/src/main/java/wannabit/io/cosmostaion/widget/mainWallet/WalletMintHolder.java
@@ -61,13 +61,11 @@ public class WalletMintHolder extends BaseHolder {
             } else {
                 msg3 = "" + WDp.getPercentDp(param.getDpRealApr(baseChain));
             }
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                msg = Html.fromHtml(mainActivity.getString(R.string.str_apr_help_onchain_msg)  + "<br>"
-                        +  msg2 + "<p>" + "<br>"
-                        + mainActivity.getString(R.string.str_apr_help_real_msg) + "<br>"
-                        +  msg3, Html.FROM_HTML_MODE_LEGACY);
-            }
-            CommonAlertDialog.showSingleButton(mainActivity, msg, null, Html.fromHtml("<font color=\"#007AFF\">" + mainActivity.getString(R.string.str_ok) + "</font>"), null);
+            msg = Html.fromHtml("<small>" + mainActivity.getString(R.string.str_apr_help_onchain_msg) + "<br/>"
+                    + msg2 + "<br/><br/>"
+                    + mainActivity.getString(R.string.str_apr_help_real_msg) + "<br/>"
+                    + msg3 + "</small>", Html.FROM_HTML_MODE_COMPACT);
+            CommonAlertDialog.showSingleButton(mainActivity, msg, null, mainActivity.getString(R.string.str_ok), null);
         });
     }
 }
