+++ app/src/main/res/layout/dialog_template_default.xml
@@ -34,6 +34,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginLeft="@dimen/space_24"
+            android:visibility="gone"
             android:layout_marginTop="@dimen/space_12"
             android:layout_marginRight="@dimen/space_24"
             android:gravity="center"
@@ -75,7 +76,6 @@
                 android:layout_weight="1"
                 android:background="@color/colorTrans"
                 android:textAllCaps="false"
-                android:visibility="gone"
                 tools:text="Button Title" />
         </LinearLayout>
     </LinearLayout>
