+++ app/src/main/res/layout/item_dialog_currency_set.xml
@@ -20,8 +20,4 @@
             android:layout_height="wrap_content"
             android:textColor="@color/colorDialog"/>
     </LinearLayout>
-<!--    <View-->
-<!--        android:layout_width="match_parent"-->
-<!--        android:layout_height="@dimen/space_view"-->
-<!--        android:background="@color/colorGray4"/>-->
 </LinearLayout>
