+++ app/src/main/res/layout/item_dialog_swap_coin.xml
@@ -1,16 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@color/colorTrans"
+    android:clickable="true"
     android:paddingLeft="@dimen/space_8"
-    android:paddingRight="@dimen/space_8"
     android:paddingTop="@dimen/space_4"
-    android:paddingBottom="@dimen/space_4"
-    android:clickable="true">
+    android:paddingRight="@dimen/space_8"
+    android:paddingBottom="@dimen/space_4">
 
     <LinearLayout
         android:id="@+id/rootLayer"
@@ -41,9 +38,8 @@
         android:id="@+id/dim_layer"
         android:layout_width="match_parent"
         android:layout_height="@dimen/space_50"
+        android:background="@drawable/box_round_gray"
         android:clickable="false"
         android:foreground="?android:attr/selectableItemBackground"
-        android:background="@drawable/box_round_gray"
-        android:visibility="gone">
-    </FrameLayout>
+        android:visibility="gone"></FrameLayout>
 </RelativeLayout>
